Refactor backend layout. Add pre_delete hook at image
