+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/parts/__init__.py
@@ -1,0 +1,4 @@
+"""
+vision_on_edge.parts init
+"""
+default_app_config = 'vision_on_edge.parts.apps.PartsConfig'
