+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/parts/admin.py
@@ -1,9 +1,9 @@
 """
-Camera admin
+Part admin
 """
 from django.contrib import admin
 
-from .models import Camera
+from .models import Part
 
 # Register your models here.
-admin.site.register(Camera)
+admin.site.register(Part)
